+++ core/internal/types/types.go
@@ -157,7 +157,7 @@ type FileUploadRequest struct {
 	Hash string `json:"hash,optional"`
 	Name string `json:"name,optional"`
 	Ext  string `json:"ext,optional"`
-	Size int32  `json:"size,optional"`
+	Size int64  `json:"size,optional"`
 	Path string `json:"path,optional"`
 }
